Extract MS2-specific methods out of ProteinManager into PeptideManager
