+++ ms2/src/org/labkey/ms2/compare/CompareQuery.java
@@ -40,6 +40,7 @@ import org.labkey.ms2.MS2Controller;
 import org.labkey.ms2.MS2Manager;
 import org.labkey.ms2.MS2Run;
 import org.labkey.ms2.MS2RunType;
+import org.labkey.ms2.PeptideManager;
 import org.labkey.ms2.protein.ProteinManager;
 import org.labkey.ms2.protein.ProteinSchema;
 import org.springframework.validation.BindException;
@@ -327,8 +328,8 @@ public class CompareQuery extends SQLFragment
 
     protected void addWhereClauses(SimpleFilter filter)
     {
-        SimpleFilter peptideFilter = ProteinManager.getPeptideFilter(_currentUrl, _runs, ProteinManager.URL_FILTER + ProteinManager.EXTRA_FILTER, _user);
-        peptideFilter = ProteinManager.reduceToValidColumns(peptideFilter, MS2Manager.getTableInfoPeptides());
+        SimpleFilter peptideFilter = PeptideManager.getPeptideFilter(_currentUrl, _runs, PeptideManager.URL_FILTER + PeptideManager.EXTRA_FILTER, _user);
+        peptideFilter = PeptideManager.reduceToValidColumns(peptideFilter, MS2Manager.getTableInfoPeptides());
         filter.addAllClauses(peptideFilter);
     }
